+++ CHANGELOG.md
@@ -1,2 +1,8 @@
 ## Changelog
-- Add new range Radar Alert for Entombed Sentinels (default disabled)
+- Fix timers for Entombed Sentinels
+- Add Pace Comparison Data for Vashnik
+- Add Setting for Anchor Position of Stack&Duration text for Aura Tracking
+- Add default disabled debuff assignment on Sszorak - Mythic only
+- Make new default disabled Soak first, second, third Alerts on Twin Fangs - Mythic only
+- Change the "Watch Spawns" Alert for tanks to "Push" and enabled for all melee instead
+- Add Pace Comparison Data for Sszorak & Twin Fangs